Complete HelloWorldStorageHandler skeleton
This commit finishes the HelloWorldStorageHandler skeleton. The
serialization and output capabilities are all left entirely unimplemented.
Deserialization and "OutputFormat" functionality is written to hardcode
a 3 item result set to any hive query on a table using this storage
handler.

The goal of this skeleton isn't to create a functioning StorageHandler,
it's just to get enough in place to reproduce the NULL table entries
described in the README (forthcoming).